+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.12.14
 
-* New option: `find_rodata_to_text`.
+* New option: `pair_rodata_to_text`.
   * If enabled, splat will try to find to which text segment an unpaired rodata segment belongs and it will hint it to the user.
 
 ### 0.12.13